// libs/ui-kit/elements/src/lib/ui-toastr/toastr-bar.enum.ts
export enum ToastrBarEventEnum {
    Created = 'created',
    Updated = 'updated',
    Deleted = 'deleted',
    ExceptionCreated = 'exceptionCreated',
    ExceptionUpdated = 'exceptionUpdated',
    ExceptionDeleted = 'exceptionDeleted',
    DeletedFile = 'deletedFile',
    DeletedFileException = 'deletedFileException',
    DownloadFileException = 'downloadFileException',
    UploadFileException = 'uploadFileException',
    UploadFile = 'uploadFile',
    Custom = 'custom',
    NotFound = 'notFound'
}